parser handles tag_details
