+++ lib/rbexy/nodes/html_node.rb
@@ -1,0 +1,6 @@
+module Rbexy
+  module Nodes
+    class HTMLNode < XMLNode
+    end
+  end
+end
